Modify ocean freezing temperature salinity dependence
This commit makes a slight modification of the calculation of the ocean
freezing temperature beneath ice shelf cavities to account for
dependence on salinity.  This adjustment matches what is currently done
in MPAS-Ocean.

Also update parameter values in shr_const_mod to match exactly what is
in MPAS-Ocean.  Descriptions of those parameters are also added.